program/requirements.txt
opencv-python
python-socketio[client]
Pillow
face-recognition
numpy
sounddevice
mediapipe

